backend/requirements_clean.txt
fastapi==0.128.7
uvicorn==0.40.0
pymongo==4.16.0
python-dotenv==1.2.1
bcrypt==5.0.0
python-jose==3.5.0
pydantic==2.12.5
transformers==4.46.3
torch==2.5.1
safetensors==0.7.0
requests==2.32.5